components/gpu_agent/requirements.txt
pydantic>=1.10,<3.0
PyYAML>=6.0
nvidia-ml-py3>=7.352.0
